+++ WebApplication/src/components/TokenLanding.vue
@@ -82,7 +82,7 @@ export default {
     },
     methods: {
         titleForTag(tag) {
-            return translateDicomTag(this.$i18n.t, this.$i18n.te, tag);
+            return translateDicomTag(tag);
         },
         valueForTag(tag) {
             if (tag in this.studyMainDicomTags) {
@@ -214,9 +214,9 @@ export default {
                         <span class="visually-hidden">Loading...</span>
                     </div>
                 </span>
-                <p v-if="tokenChecked && errorCode == 'invalid'" v-html="$t('token.error_token_invalid_html')"></p>
-                <p v-if="tokenChecked && errorCode == 'expired'" v-html="$t('token.error_token_expired_html')"></p>
-                <p v-if="tokenChecked && errorCode == 'unknown'" v-html="$t('token.error_token_unknown_html')"></p>
+                <p v-if="tokenChecked && errorCode == 'invalid'" style="white-space: pre-line" v-html="$t('token.error_token_invalid_html')"></p>
+                <p v-if="tokenChecked && errorCode == 'expired'" style="white-space: pre-line" v-html="$t('token.error_token_expired_html')"></p>
+                <p v-if="tokenChecked && errorCode == 'unknown'" style="white-space: pre-line" v-html="$t('token.error_token_unknown_html')"></p>
             </div>
             <div v-if="tokenChecked && showOptions" class="w-100">
                 <div v-if="showStudyDetails" class="w-100">